Related with issue 27712: Fixes API Change for NOT NULL column M_ATTRIBUTESETINSTANCE_ID
There is no risk in setting it to NOT NULL as  a default an onCreateDefault have been added. This column was already being populated in most of the cases with '0' as default value. It was forcing many sql to use coalesce while just few records were null (although many already contained '0')