Fixes issue 28847: Fixes API change (added final modifier)
In this changeset [1] the final modifier has been added to a method. This has broken the API. There is no need to notify this API change because the method that caused was created in a version that has not been released yet. In the previous version the wrong version of java.japi.gz was included.

[1] https://code.openbravo.com/erp/devel/pi/rev/128a5d39aba2a3e2e979a54f143c31d8fbf83b65