Fixed bug 33463: Java (from xsql) API change related to issue 33460
The xsql query in charge of getting the data now has a new parameter to control whether to query or not for the comparative year. Thus we can improve the performance of the query, avoiding to retrieve data that is not going to be shown later on in the report.

This xsql file builds a public Java class and method. The method now has a new parameter which creates the API change.